+++ charm/Mcpd8.CmdPacket.cpp
@@ -1,7 +1,7 @@
 #include "Mcpd8.CmdPacket.hpp"
 
 namespace Mcpd8 {
-		const int CmdPacket::defaultLength = 10;
+		const uint16_t CmdPacket::defaultLength = 10;
 		unsigned short sendcounter = 0;
 
 }
